refactor for util functions
