+++ package.json
@@ -5,10 +5,8 @@
 	"main": "index.js",
 	"scripts": {
 		"start": "NODE_ENV=production node server.js",
-		"dev:start":
-			"NODE_ENV=development webpack-dev-server --config webpack.config.dev.babel.js --progress",
-		"build":
-			"NODE_ENV=production webpack --config webpack.config.prod.babel.js",
+		"dev:start": "NODE_ENV=development webpack-dev-server --config webpack.config.dev.babel.js --progress",
+		"build": "NODE_ENV=production webpack --config webpack.config.prod.babel.js",
 		"prebuild": "echo \"getting ready to build\" && rm -rf dist/",
 		"test": "echo \"Error: no test specified\" && exit 0",
 		"precommit": "npm run lint && npm run flow",
@@ -88,7 +86,6 @@
 		"react-router": "^4.2.0",
 		"react-router-dom": "^4.2.2",
 		"recompose": "^0.26.0",
-		"startwars-name": "^1.1.0",
 		"styled-components": "^2.1.2",
 		"url-loader": "^0.6.2",
 		"webpack-hot-middleware": "^2.20.0"
