+++ src/utils/starwars.json
@@ -1,0 +1,12 @@
+[
+	"Luke Skywalker",
+	"C-3PO",
+	"R2-D2",
+	"Darth Vader",
+	"Leia Organa",
+	"Owen Lars",
+	"Beru Whitesun lars",
+	"R5-D4",
+	"Biggs Darklighter",
+	"Obi-Wan Kenobi"
+]
